% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colegio.R
\docType{data}
\name{colegio}
\alias{colegio}
\title{Datos del colegio Inem}
\format{
\subsection{\code{colegio}}{

A data frame with 3 rows and 3 columns:
\describe{
\item{edad}{edad del estudiante en años}
\item{peso}{peso corporal en kg}
\item{gafas}{es TRUE si el estudiante usa gafas, caso contrario FALSE}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
colegio
}
\description{
Este conjunto de datos pertenece a estudiantes de 6 grado del colegio Inem
}
\keyword{datasets}
