% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promedio.R
\name{promedio}
\alias{promedio}
\title{Función para calcular el promedio de un vector}
\usage{
promedio(x)
}
\arguments{
\item{x}{vector numérico}
}
\value{
Esta función retorna el valor del promedio de un vector.
}
\description{
Función para calcular el promedio de un vector
}
\examples{
# Ejemplo 1
promedio(x=c(3, 6, 8, 0, 1))
# Ejemplo 2
y <- runif(n=30, min=-5, max=3)
promedio(x=y)
}
