% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Función para sumar}
\usage{
suma(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo número}
}
\value{
El resultado de la suma
}
\description{
Función para sumar
}
\examples{
suma(a=4, b=6)
suma(b=7, a=3)
}
