% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_tres.R
\name{suma_tres}
\alias{suma_tres}
\alias{multi_tres}
\title{Esta función opera tres valores}
\usage{
suma_tres(a, b, c)

multi_tres(a, b, c)
}
\arguments{
\item{a}{primer valor que se ingresa.}

\item{b}{segundo valor que se ingresa.}

\item{c}{tercer valor que se ingresa.}
}
\value{
Esta función retorna un valor con el resultado.
}
\description{
Esta función opera tres valores
}
\examples{
suma_tres(a=3, b=5, c=2)
multi_tres(a=9, b=0,c=-2)
}
