% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promedio.R
\name{promedio}
\alias{promedio}
\title{Funcion para calcular el promedio de un vector}
\usage{
promedio(x)
}
\arguments{
\item{x}{vector numerico.}
}
\value{
esta funcion retorna el valor del promedio de un vector.
}
\description{
Funcion para calcular el promedio de un vector
}
\examples{
# Ejemplo 1
promedio(x=c(3, 6, 8, 0, 1))
# Ejemplo 2
y <- runif(n=30, min=-5, max=3)
promedio(x=y)
}
